+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/Cards/ODSCardImageFirst.swift
@@ -189,10 +189,10 @@ struct ODSCardImageFirst_Previews: PreviewProvider {
     }
 
     static let model = ODSCardImageFirstModel(
-        title: ODSCardModel.example.title,
-        subtitle: ODSCardModel.example.subTitle,
-        image: Image("ods_empty", bundle: Bundle.ods),
-        supportingText: ODSCardModel.example.description)
+        title: ODSCCardPreviewData.title,
+        subtitle: ODSCCardPreviewData.subtitle,
+        image: ODSCCardPreviewData.image,
+        supportingText: ODSCCardPreviewData.supportingText)
 
     struct TestView: View {
         @State var showTextInTost: String?