+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - [DemoApp/SDK] #222 Update the Button page to list the available button types
 - [SDK] Bug #96 Add Small Card 
+- [SDK/DemoApp] #229 Technical update to use Small and ImageFirst cards in demo app and modules 
 - [DemoApp] #228 Change Chip to switch on List screen for "SecondLineOfText" and "LeadingImage" options 
 - [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components
 - [SDK/DemoApp] #182 Add card title first